Move tests to agb_alloc mod.rs
